Start splitting protections
